Cleaned up train.py a bit, added dataset loading functions
